FOCUSING User case list
Use case list for DTALite's new release.

F - Fundamental Use Cases:
1. UE as Fundamental: Calculate link, route, OD, and system performance using default demand for a given network.
2. 2-corridor User Equilibrium: Determine route, link, and OD with a demand volume of 7000.

+++ src/cpp/VDF.h
@@ -451,7 +451,7 @@ public:
            return avg_travel_time;
      }
 
-     std::map<int, int> turn_link_count_map;
+     std::map<int, float> turn_link_count_map;
      std::map<int, int> restricted_turn_nodes_map;
      string restricted_turn_nodes_str; 
      e_VDF_type vdf_type;
